imp(ante): refactor AnteHandler (#1479)
* imp(ante): refactor AnteHandler

* fix test

* test

* Adjust deprecated sdkerrors import (#1493)

* refactor test files

* Apply suggestions from code review


* lint

* prioritization comment

* fix test
